# api/openapi.runtime.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JobPilot Platform API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/applications": {
   "get": {
    "tags": [
     "applications"
    ],
    "summary": "List Applications",
    "operationId": "list_applications_api_v1_applications_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 25,
       "title": "Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApplicationPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "applications"
    ],
    "summary": "Create Application",
    "operationId": "create_application_api_v1_applications_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateApplicationIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApplicationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/applications/{application_id}": {
   "get": {
    "tags": [
     "applications"
    ],
    "summary": "Get Application",
    "operationId": "get_application_api_v1_applications__application_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Application Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApplicationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "applications"
    ],
    "summary": "Update Application",
    "operationId": "update_application_api_v1_applications__application_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Application Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdateApplicationIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApplicationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "applications"
    ],
    "summary": "Delete Application",
    "operationId": "delete_application_api_v1_applications__application_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Application Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/applications/{application_id}/status": {
   "post": {
    "tags": [
     "applications"
    ],
    "summary": "Change Status",
    "operationId": "change_status_api_v1_applications__application_id__status_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Application Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChangeStatusIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApplicationOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/applications/{application_id}/events": {
   "get": {
    "tags": [
     "applications"
    ],
    "summary": "List Events",
    "operationId": "list_events_api_v1_applications__application_id__events_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Application Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ApplicationEventOut"
         },
         "title": "Response List Events Api V1 Applications  Application Id  Events Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/applications/{application_id}/apply": {
   "post": {
    "tags": [
     "applications"
    ],
    "summary": "Auto Apply",
    "description": "Compliance gate for automated application.\n\nAutomation is only ever permitted for OFFICIAL_API /\nUSER_AUTHORIZED_AUTOMATION connectors that opt in \u2014 everything else gets\na 409 with the manual-apply URL. Permitted sources return 501 until the\nper-ATS submitter ships; this endpoint never fakes a submission.",
    "operationId": "auto_apply_api_v1_applications__application_id__apply_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Application Id"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "ApplicationEventOut": {
    "properties": {
     "from_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "From Status"
     },
     "to_status": {
      "type": "string",
      "title": "To Status"
     },
     "actor": {
      "type": "string",
      "title": "Actor"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "to_status",
     "actor",
     "created_at"
    ],
    "title": "ApplicationEventOut"
   },
   "ApplicationOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "job": {
      "$ref": "#/components/schemas/JobOut"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "method": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Method"
     },
     "applied_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Applied At"
     },
     "deadline_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline At"
     },
     "next_action_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Action At"
     },
     "salary_offered": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Offered"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "job",
     "status",
     "created_at",
     "updated_at"
    ],
    "title": "ApplicationOut"
   },
   "ApplicationPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ApplicationOut"
      },
      "type": "array",
      "title": "Items"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "size": {
      "type": "integer",
      "title": "Size"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "items",
     "page",
     "size",
     "total"
    ],
    "title": "ApplicationPage"
   },
...
   "ChangeStatusIn": {
    "properties": {
     "to_status": {
      "type": "string",
      "enum": [
       "SAVED",
       "INTERESTED",
       "RESUME_GENERATED",
       "APPLIED",
       "RECRUITER_CONTACTED",
       "OA_RECEIVED",
       "INTERVIEW_SCHEDULED",
       "REJECTED",
       "OFFER",
       "ACCEPTED",
       "DECLINED"
      ],
      "title": "To Status"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "type": "object",
    "required": [
     "to_status"
    ],
    "title": "ChangeStatusIn"
   },
...
   "CreateApplicationIn": {
    "properties": {
     "job_id": {
      "type": "string",
      "format": "uuid",
      "title": "Job Id"
     },
     "status": {
      "type": "string",
      "enum": [
       "SAVED",
       "INTERESTED",
       "RESUME_GENERATED",
       "APPLIED",
       "RECRUITER_CONTACTED",
       "OA_RECEIVED",
       "INTERVIEW_SCHEDULED",
       "REJECTED",
       "OFFER",
       "ACCEPTED",
       "DECLINED"
      ],
      "title": "Status",
      "default": "SAVED"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "job_id"
    ],
    "title": "CreateApplicationIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "JobOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "connector_id": {
      "type": "string",
      "title": "Connector Id"
     },
     "url": {
      "type": "string",
      "title": "Url"
     },
     "location_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location Text"
     },
     "workplace": {
      "type": "string",
      "title": "Workplace"
     },
     "employment": {
      "type": "string",
      "title": "Employment"
     },
     "arrangement": {
      "type": "string",
      "title": "Arrangement"
     },
     "salary_min": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Min"
     },
     "salary_max": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Max"
     },
     "salary_currency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Currency"
     },
     "posted_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Posted At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "title",
     "connector_id",
     "url",
     "workplace",
     "employment",
     "arrangement"
    ],
    "title": "JobOut"
   },
...
   "UpdateApplicationIn": {
    "properties": {
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "salary_offered": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Salary Offered"
     },
     "deadline_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline At"
     },
     "next_action_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Action At"
     },
     "resume_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Resume Id"
     }
    },
    "type": "object",
    "title": "UpdateApplicationIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}